requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.1.10
django-cors-headers==4.2.0
django-imagefield==0.16.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
dnspython==2.4.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
httplib2==0.22.0
idna==3.4
oauth2client==4.1.3
oauthlib==3.2.2
Pillow==10.0.0
postgres==4.0
psycopg2-binary==2.9.7
psycopg2-pool==1.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pymongo==4.4.1
pyparsing==3.1.1
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.2.4
tzdata==2023.3
urllib3==1.26.16
